git-svn-id: https://repo.acin.tuwien.ac.at/v4r/stable/strandsv4r@6806 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ v4r/SurfaceRelations/BoundaryRelationsBase.cpp
@@ -72,7 +72,7 @@ void BoundaryRelationsBase::setNormals(pcl::PointCloud<pcl::Normal>::Ptr _normal
   if(!have_cloud)
     throw std::runtime_error("[BoundaryRelationsBase::setNormals] I suggest you first set the point cloud.");
   
-  if ( (_normals->height!=height) || (_normals->width!=width) )
+  if ( (_normals->height!=(unsigned int)height) || (_normals->width!=(unsigned int)width) )
     throw std::runtime_error("[BoundaryRelationsBase::setNormals] Invalid normals (not for this point cloud).");
 
   normals = _normals;
